+++ test/kino_wardley.test.js
@@ -15,6 +15,29 @@ test('Handle inertia', () => {
          "dx": -27,
          "dy": -12,
          "inertia": true,
+         "xor": false,
+         "name": "Customer",
+         "type": "component",
+         "x": 0.5,
+         "y": 0.95,
+       },
+     }
+
+  expect(parseSpec(input.split("\n"))).toStrictEqual(expected);
+})
+
+test('Handle xor', () => {
+  const input = `
+  component Customer [0.95, 0.5] xor label [-27, -12]
+  `
+
+  const expected =    
+      {
+       "Customer":  {
+         "dx": -27,
+         "dy": -12,
+         "inertia": false,
+         "xor": true,
          "name": "Customer",
          "type": "component",
          "x": 0.5,
@@ -64,6 +87,7 @@ test('Sample spec parses to expected', () => {
       "name": "Cup",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": 0,
       "dy": 0,
       "x": 0.78,
@@ -73,6 +97,7 @@ test('Sample spec parses to expected', () => {
       "name": "Cup of Tea",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": 0,
       "dy": 0,
       "x": 0.61,
@@ -82,6 +107,7 @@ test('Sample spec parses to expected', () => {
       "name": "Hot Water",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": 0,
       "dy": 0,
       "x": 0.8,
@@ -91,6 +117,7 @@ test('Sample spec parses to expected', () => {
       "name": "Kettle",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": 0,
       "dy": 0,
       "x": 0.35,
@@ -108,6 +135,7 @@ test('Sample spec parses to expected', () => {
       "name": "Power",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": 0,
       "dy": 0,
       "x": 0.71,
@@ -133,6 +161,7 @@ test('Sample spec parses to expected', () => {
       "name": "Tea",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": 0,
       "dy": 0,
       "x": 0.81,
@@ -142,6 +171,7 @@ test('Sample spec parses to expected', () => {
       "name": "Water",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": 0,
       "dy": 0,
       "x": 0.82,
@@ -204,6 +234,7 @@ test('Online wardley map parses to expected', () => {
       "name": "Cup",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": 0,
       "dy": 0,
       "x": 0.78,
@@ -213,6 +244,7 @@ test('Online wardley map parses to expected', () => {
       "name": "Cup of Tea",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": 19,
       "dy": -4,
       "x": 0.61,
@@ -222,6 +254,7 @@ test('Online wardley map parses to expected', () => {
       "name": "Hot Water",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": 0,
       "dy": 0,
       "x": 0.8,
@@ -231,6 +264,7 @@ test('Online wardley map parses to expected', () => {
       "name": "Kettle",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": -57,
       "dy": 4,
       "x": 0.35,
@@ -248,6 +282,7 @@ test('Online wardley map parses to expected', () => {
       "name": "Power",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": -27,
       "dy": 20,
       "x": 0.7,
@@ -273,6 +308,7 @@ test('Online wardley map parses to expected', () => {
       "name": "Tea",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": 0,
       "dy": 0,
       "x": 0.81,
@@ -282,6 +318,7 @@ test('Online wardley map parses to expected', () => {
       "name": "Water",
       "type": "component",
       "inertia": false,
+      "xor": false,
       "dx": 0,
       "dy": 0,
       "x": 0.82,