+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Pending
+## [0.7.0]
 
 ### Added inertia to Components
 ### Added some new Wardley Map components based upon a tweet (https://twitter.com/swardley/status/1565744194475499521)
+### Added Market
+### Fixed documentation of labels.
 
 ## [0.6.0]